Major Tags refactor
