+++ lib/solid/parser/argument.ex
@@ -40,5 +40,22 @@ defmodule Solid.Parser.Argument do
     |> tag(:named_arguments)
   end
 
+  def filter() do
+    filter_name =
+      ascii_string([?a..?z, ?A..?Z], 1)
+      |> concat(ascii_string([?a..?z, ?A..?Z, ?_], min: 0))
+      |> reduce({Enum, :join, []})
+
+    ignore(space())
+    |> ignore(string("|"))
+    |> ignore(space())
+    |> concat(filter_name)
+    |> tag(
+      optional(ignore(string(":")) |> ignore(space()) |> concat(arguments())),
+      :arguments
+    )
+    |> tag(:filter)
+  end
+
   def arguments(), do: choice([named_arguments(), positional_arguments()])
 end
