+++ test/integration/custom_tag_test.exs
@@ -4,72 +4,12 @@ defmodule Solid.Integration.CustomTagsTest do
 
   describe "custom tags" do
     test "pass in custom tag that needs no arguments" do
-      assert render("{% current_date %}", %{},
-               tags: %{"current_date" => CustomTags.CurrentDate},
-               parser: CustomDateParser
-             ) ==
+      assert render("{% get_current_date %}", %{}, parser: CustomDateParser) ==
                to_string(DateTime.utc_now().year)
     end
 
     test "pass in custom tag that needs arguments" do
-      assert render(~s({% get_year_of_date "2020-08-06T06:23:48Z" %}), %{},
-               tags: %{"get_year_of_date" => CustomTags.GetYearOfDate},
-               parser: CustomDateParser
-             ) ==
-               "2020-8-6"
-    end
-
-    test "custom tags are evaluated inside of for loops" do
-      assert render(
-               "{% for date in dates %}<span>{% get_year_of_date date %}</span>{% endfor %}",
-               %{"dates" => ["2020-08-06T06:23:48Z", "2021-05-03T06:23:48Z"]},
-               tags: %{"get_year_of_date" => CustomTags.GetYearOfDate},
-               parser: CustomDateParser
-             ) == "<span>2020-8-6</span><span>2021-5-3</span>"
-    end
-
-    test "custom tags are evaluated inside of captures" do
-      assert render(
-               ~s({% capture testing %}{% get_year_of_date "2020-08-06T06:23:48Z" %}{% endcapture %}{{ testing }}),
-               %{"dates" => ["2020-08-06T06:23:48Z", "2021-05-03T06:23:48Z"]},
-               tags: %{"get_year_of_date" => CustomTags.GetYearOfDate},
-               parser: CustomDateParser
-             ) == "2020-8-6"
-    end
-
-    test "custom tags are evaluated inside of for loops and captures" do
-      rendered =
-        render(
-          """
-          {% capture testing %}
-            {% for date in dates %}<span>{% get_year_of_date date %}</span>{% endfor %}
-          {% endcapture %}
-
-          {{ testing }}
-          """,
-          %{"dates" => ["2020-08-06T06:23:48Z", "2021-05-03T06:23:48Z"]},
-          tags: %{"get_year_of_date" => CustomTags.GetYearOfDate},
-          parser: CustomDateParser
-        )
-
-      assert String.trim(rendered) == "<span>2020-8-6</span><span>2021-5-3</span>"
-    end
-  end
-
-  describe "custom tags modules" do
-    test "pass in custom tag that needs no arguments" do
-      assert render("{% get_current_date %}", %{},
-               tags: %{"get_current_date" => CustomTags.CurrentDate},
-               parser: CustomDateParser
-             ) ==
-               to_string(DateTime.utc_now().year)
-    end
-
-    test "pass in custom tag that needs arguments" do
-      assert render(~s({% get_year "2020-08-06T06:23:48Z" %}), %{},
-               tags: %{"get_year" => CustomTags.GetYearOfDate},
-               parser: CustomDateParser
-             ) ==
+      assert render(~s({% get_year "2020-08-06T06:23:48Z" %}), %{}, parser: CustomDateParser) ==
                "2020-8-6"
     end
 
@@ -77,7 +17,6 @@ defmodule Solid.Integration.CustomTagsTest do
       assert render(
                "{% for date in dates %}<span>{% get_year date %}</span>{% endfor %}",
                %{"dates" => ["2020-08-06T06:23:48Z", "2021-05-03T06:23:48Z"]},
-               tags: %{"get_year" => CustomTags.GetYearOfDate},
                parser: CustomDateParser
              ) == "<span>2020-8-6</span><span>2021-5-3</span>"
     end
@@ -86,7 +25,6 @@ defmodule Solid.Integration.CustomTagsTest do
       assert render(
                ~s({% capture testing %}{% get_year "2020-08-06T06:23:48Z" %}{% endcapture %}{{ testing }}),
                %{"dates" => ["2020-08-06T06:23:48Z", "2021-05-03T06:23:48Z"]},
-               tags: %{"get_year" => CustomTags.GetYearOfDate},
                parser: CustomDateParser
              ) == "2020-8-6"
     end
@@ -102,7 +40,6 @@ defmodule Solid.Integration.CustomTagsTest do
           {{ testing }}
           """,
           %{"dates" => ["2020-08-06T06:23:48Z", "2021-05-03T06:23:48Z"]},
-          tags: %{"get_year" => CustomTags.GetYearOfDate},
           parser: CustomDateParser
         )
 
@@ -118,7 +55,6 @@ defmodule Solid.Integration.CustomTagsTest do
           {% endmyblock %}
           """,
           %{},
-          tags: %{"myblock" => CustomTags.CustomBrackedWrappedTag},
           parser: CustomDateParser
         )
