+++ test/integration/tags_test.exs
@@ -331,26 +331,15 @@ defmodule Solid.Integration.TagsTest do
   end
 
   describe "custom tag" do
-    defmodule FoobarTag do
-      def render(_context, _arguments) do
-        "barbaz"
-      end
-    end
-
-    defmodule FoobarValTag do
-      def render(_context, arguments: [value: string]) do
-        "barbaz#{string}"
-      end
-    end
-
     test "with no arguments" do
       text = """
       {% foobar %}
       """
 
-      assert render(text, %{}, tags: %{"foobar" => FoobarTag}, parser: CustomFooParser) == """
-             barbaz
-             """
+      assert render(text, %{}, parser: CustomFooParser) ==
+               """
+               barbaz
+               """
     end
 
     test "with an argument" do
@@ -358,7 +347,7 @@ defmodule Solid.Integration.TagsTest do
       {% foobarval "-show-me" %}
       """
 
-      assert render(text, %{}, tags: %{"foobarval" => FoobarValTag}, parser: CustomFooParser) ==
+      assert render(text, %{}, parser: CustomFooParser) ==
                """
                barbaz-show-me
                """
