+++ test/support/custom_parsers.ex
@@ -1,14 +1,16 @@
 defmodule CustomDateParser do
   use Solid.Parser.Base,
     custom_tags: [
-      "current_date",
-      "get_year_of_date",
-      {"get_current_date", CustomTags.CurrentDate},
-      {"get_year", CustomTags.GetYearOfDate},
-      {"myblock", CustomTags.CustomBrackedWrappedTag}
+      get_current_date: CustomTags.CurrentDate,
+      get_year: CustomTags.GetYearOfDate,
+      myblock: CustomTags.CustomBrackedWrappedTag
     ]
 end
 
 defmodule CustomFooParser do
-  use Solid.Parser.Base, custom_tags: ["foobar", "foobarval"]
+  use Solid.Parser.Base,
+    custom_tags: [
+      foobar: CustomTags.FoobarTag,
+      foobarval: CustomTags.FoobarValTag
+    ]
 end
